one more fixed PR comments, methods moved to another class
