+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/db/DbUtils.java
@@ -61,37 +61,6 @@ public final class DbUtils {
         }
     }
 
-    /**
-     * Close a reader without throwing an exception.
-     *
-     * @param reader the reader or null
-     */
-    public static void closeSilently(Reader reader) {
-        if (reader != null) {
-            try {
-                reader.close();
-            } catch (Exception e) {
-                // ignore
-            }
-        }
-    }
-
-    /**
-     * Close a writer without throwing an exception.
-     *
-     * @param writer the writer or null
-     */
-    public static void closeSilently(Writer writer) {
-        if (writer != null) {
-            try {
-                writer.flush();
-                writer.close();
-            } catch (Exception e) {
-                // ignore
-            }
-        }
-    }
-
     public static void rollback(Connection con) {
         try {
             if (con != null) {
